# employee-api/openapi.yaml
openapi: 3.0.1
info:
  title: Mock Employee API
  version: 1.0.0
  description: A simple mock employee API service
servers:
  - url: http://localhost:4000
    description: Local server

paths:
  /employees:
    get:
      summary: Get all employees
      responses:
        '200':
          description: A list of employees
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  $ref: '#/components/schemas/Employee'
    post:
      summary: Create a new employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Employee'
      responses:
        '201':
          description: Employee created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'

  /employees/{id}:
    get:
      summary: Get a specific employee by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: The employee ID
      responses:
        '200':
          description: The employee data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '404':
          description: Employee not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Update an existing employee
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: The employee ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Employee'
      responses:
        '200':
          description: Employee updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '404':
          description: Employee not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete an employee
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: The employee ID
      responses:
        '204':
          description: Employee deleted
        '404':
          description: Employee not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    Employee:
      type: object
      properties:
        name:
          type: string
        position:
          type: string
    Error:
      type: object
      properties:
        error:
          type: string